requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
scikit-image>=0.21.0

# Experiment tracking
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
omegaconf>=2.3.0
safetensors>=0.4.0
einops>=0.7.0
timm>=0.9.0
requests>=2.28.0
packaging>=21.0

# Web interfaces
gradio>=3.50.0
streamlit>=1.28.0

# Development tools
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0

# COCO dataset support
pycocotools>=2.0.6
